DRY up bundler/capture testing utilities
