+++ test/tomo/cli/completions_test.rb
@@ -1,14 +1,10 @@
 require "test_helper"
-require "fileutils"
-require "open3"
-require "securerandom"
-require "tmpdir"
 
 class Tomo::CLI::CompletionsTest < Minitest::Test
   def test_completions_include_setting_names
-    output = in_temp_dir do
-      capture! "bundle exec tomo init"
-      capture! "bundle exec tomo --complete deploy -s"
+    output = Tomo::Testing::Local.in_temp_dir do
+      capture "bundle exec tomo init"
+      capture "bundle exec tomo --complete deploy -s"
     end
 
     assert_match(/^git_branch=$/, output)
@@ -17,19 +13,9 @@ class Tomo::CLI::CompletionsTest < Minitest::Test
 
   private
 
-  def in_temp_dir(&block)
-    dir = File.join(Dir.tmpdir, "tomo_test_#{SecureRandom.hex(8)}")
-    FileUtils.mkdir_p(dir)
-    Dir.chdir(dir, &block)
-  end
-
-  def capture!(command)
-    Bundler.with_original_env do
-      gemfile = File.expand_path("../../../Gemfile", __dir__)
-      output, status = Open3.capture2({ "BUNDLE_GEMFILE" => gemfile }, command)
-      raise "Command failed: #{command}" unless status.success?
-
-      output
+  def capture(command)
+    Tomo::Testing::Local.with_tomo_gemfile do
+      Tomo::Testing::Local.capture(command)
     end
   end
 end
